+++ crates/openvcs-core/src/lib.rs
@@ -94,12 +94,38 @@ pub trait Vcs: Send + Sync {
 pub struct Repo {
     inner: Arc<dyn Vcs>,
 }
+
 impl Repo {
-    pub fn new(inner: Arc<dyn Vcs>) -> Self { Self { inner } }
-    pub fn id(&self) -> BackendId { self.inner.id() }
-    pub fn caps(&self) -> Capabilities { self.inner.caps() }
-    pub fn inner(&self) -> &dyn Vcs { &*self.inner }
+    pub fn new(inner: Arc<dyn Vcs>) -> Self {
+        // Log once on construction so we know which backend we wrapped.
+        log::debug!("openvcs-core: Repo created with backend {}", inner.id());
+        Self { inner }
+    }
+
+    #[inline]
+    pub fn id(&self) -> BackendId {
+        let id = self.inner.id();
+        log::trace!("openvcs-core: Repo::id -> {id}");
+        id
+    }
+
+    #[inline]
+    pub fn caps(&self) -> Capabilities {
+        let caps = self.inner.caps();
+        log::trace!(
+            "openvcs-core: Repo::caps -> commits={}, branches={}, tags={}, staging={}, push_pull={}, fast_forward={}",
+            caps.commits, caps.branches, caps.tags, caps.staging, caps.push_pull, caps.fast_forward
+        );
+        caps
+    }
+
+    #[inline]
+    pub fn inner(&self) -> &dyn Vcs {
+        log::trace!("openvcs-core: Repo::inner");
+        &*self.inner
+    }
 }
+
 
 /* ========================= Runtime backend registry =========================
    Backends contribute a `BackendDescriptor` into the distributed slice below.
@@ -121,13 +147,38 @@ pub static BACKENDS: [BackendDescriptor] = [..];
 
 /// Enumerate all registered backends (order is link-order; do not rely on it).
 pub fn list_backends() -> impl Iterator<Item = &'static BackendDescriptor> {
-    BACKENDS.iter()
+    use log::{debug, trace};
+
+    // Create the iterator first so we can both inspect and return it.
+    let it = BACKENDS.iter();
+
+    // Cheap to ask the length from the slice iterator.
+    debug!("openvcs-core: {} backends registered", it.len());
+
+    // Optionally enumerate each backend at trace level.
+    for b in it.clone() {
+        trace!("openvcs-core: backend loaded: {} ({})", b.id, b.name);
+    }
+
+    it
 }
 
 /// Lookup a backend descriptor by id.
 pub fn get_backend(id: &str) -> Option<&'static BackendDescriptor> {
-    BACKENDS.iter().find(|b| b.id == id)
+    use log::{debug, warn};
+
+    match BACKENDS.iter().find(|b| b.id == id) {
+        Some(b) => {
+            debug!("openvcs-core: backend lookup succeeded → {} ({})", b.id, b.name);
+            Some(b)
+        }
+        None => {
+            warn!("openvcs-core: backend lookup failed for id='{id}'");
+            None
+        }
+    }
 }
+
 
 /* ================================ DTOs ===================================== */
