+++ crates/openvcs-git-libgit2/Cargo.toml
@@ -9,3 +9,4 @@ git2 = "0.20"
 thiserror = "2"
 serde = "1.0.219"
 linkme = "0.3"
+log = "0.4"
