+++ crates/openvcs-git/Cargo.toml
@@ -6,3 +6,4 @@ edition = "2024"
 [dependencies]
 openvcs-core = { path = "../openvcs-core" }
 linkme = "0.3"
+log = "0.4"
