+++ src-tauri/Cargo.toml
@@ -41,4 +41,6 @@ tokio = "1.47.1"
 dirs = "6.0.0"
 regex = "1.11.2"
 chrono = "0.4.41"
+log = "0.4" 
+env_logger = "0.11.8"
